Minor wording changes
Thanks Dan for this great write-up! I made some minor changes. Though everything you wrote in the motivation is correct, we want to focus the students on how the data is used for social good, so I made some modifications to that. 

Also, students won't actually completely build the app, so I made some modifications to that as well. 

I'd imagine that this would be the ideation stage in Week 1. Were you planning on having everything in one document that they use throughout the 4 weeks? Or were you going to have one for each week?